Reorganize utils file
